refactor: 로또우승객체 제거
